Move group-by types and utils to common API layer
Moves getGroupByExpression, getGroupByOrderExpression, and their
supporting types/utils from the GraphQL layer into the common
query-runner layer where they are actually consumed. Inlines the
IANA timezone validation directly into getGroupByExpression and
removes the separate sql-sanitization folder.
