+++ packages/twenty-server/src/engine/api/common/common-query-runners/types/group-by-definition.type.ts
