+++ packages/twenty-server/src/engine/api/common/common-query-runners/utils/is-group-by-relation-field.util.ts
@@ -1,7 +1,7 @@
 import {
   type GroupByField,
   type GroupByRelationField,
-} from 'src/engine/api/graphql/graphql-query-runner/group-by/resolvers/types/group-by-field.types';
+} from 'src/engine/api/common/common-query-runners/types/group-by-field.types';
 
 export const isGroupByRelationField = (
   groupByField: GroupByField,
