+++ packages/twenty-server/src/engine/api/common/common-query-runners/utils/remove-quote.util.ts
